Fixed the issue with --watch and on_demand
When there was a "watcher" it would also add an on_demand listender which would catch the messages preventing the real endpoint from receiving the message and firing.
We now only create a lisenter when something tries to listen (i.e. an endpoint). The config watcher never actually runs anything and therefor never "listens".