Use clap derive, fixing behavior of `--include` flag. (#121)
* Add clap derive macros and attributes to structs

* restore default behavior of stats_file

* Use static Lazy for a Regex

* Use Tmp struct for try config, needed for results_dir

* comments

* fix typo in comment

* another typo

* Direct port of tests from builder to derive version.

Some tests now fail as a result of the new `-i` flag behavior.

* Fix broken tests to use new `-i` behavior.

* Add new test that uses built-in clap checker

* Add new test.

Verifies that the config file can come after an --include flag, and checks that a Ne filter can be properly parsed.

* Remove old code, update use paths.

* New test: checks that old `-i` behavior doesn't work

* Improve consistency between output depicted in book and actual output.

The text wrapping of the option descriptions depends on the size of the terminal window when the program is run. A wrapping length that is relatively short but leaves the descriptions in a fully separate column was chosen for the example in the book.